+++ examples-extra/basic/cli/package.json
@@ -31,7 +31,7 @@
     "@osdk/api": "workspace:^",
     "@osdk/client": "workspace:^",
     "@osdk/examples.basic.sdk": "workspace:^",
-    "@osdk/foundry": "workspace:^",
+    "@osdk/internal.foundry": "workspace:^",
     "pino": "^8.20.0",
     "tiny-invariant": "^1.3.1"
   },
