+++ packages/client/src/object/aggregate.test.ts
@@ -15,7 +15,7 @@
  */
 
 import type { ObjectTypeDefinition, OntologyDefinition } from "@osdk/api";
-import type { AggregateObjectsResponseV2 } from "@osdk/foundry";
+import type { AggregateObjectsResponseV2 } from "@osdk/internal.foundry";
 import type { TypeOf } from "ts-expect";
 import { expectType } from "ts-expect";
 import { describe, expectTypeOf, it, type Mock, vi } from "vitest";
