+++ packages/client/src/objectSet/toConjureObjectSet.ts
@@ -15,7 +15,10 @@
  */
 
 import type { ObjectTypeKeysFrom, OntologyDefinition } from "@osdk/api";
-import type { ObjectSet, SearchJsonQueryV2 } from "@osdk/foundry/types";
+import type {
+  ObjectSet,
+  SearchJsonQueryV2,
+} from "@osdk/internal.foundry/types";
 import type {
   ObjectSet as OssObjectSet,
   ObjectSetFilter,
