+++ packages/client/src/ontology/StandardOntologyProvider.ts
@@ -18,16 +18,16 @@ import type {
   ObjectOrInterfaceDefinition,
   ObjectTypeDefinition,
 } from "@osdk/api";
-import { listInterfaceTypes } from "@osdk/foundry/OntologiesV2_OntologyObjectV2";
-import { getOntologyFullMetadata } from "@osdk/foundry/OntologiesV2_OntologyV2";
-import type {
-  ListInterfaceTypesResponse,
-  OntologyFullMetadata,
-} from "@osdk/foundry/types";
 import {
   __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition,
   wireObjectTypeFullMetadataToSdkObjectTypeDefinition,
 } from "@osdk/generator-converters";
+import { listInterfaceTypes } from "@osdk/internal.foundry/OntologiesV2_OntologyObjectV2";
+import { getOntologyFullMetadata } from "@osdk/internal.foundry/OntologiesV2_OntologyV2";
+import type {
+  ListInterfaceTypesResponse,
+  OntologyFullMetadata,
+} from "@osdk/internal.foundry/types";
 import deepEqual from "fast-deep-equal";
 import type { MinimalClient } from "../MinimalClientContext.js";
 import { createAsyncCache } from "../object/Cache.js";
