+++ packages/client/src/ontology/loadInterfaceDefinition.ts
@@ -15,8 +15,8 @@
  */
 
 import type { InterfaceDefinition } from "@osdk/api";
-import { getInterfaceType } from "@osdk/foundry/OntologiesV2_OntologyObjectV2";
 import { __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition } from "@osdk/generator-converters";
+import { getInterfaceType } from "@osdk/internal.foundry/OntologiesV2_OntologyObjectV2";
 import type { MinimalClient } from "../MinimalClientContext.js";
 
 export async function loadInterfaceDefinition(
