+++ packages/internal.foundry/README.md
@@ -1,0 +1,9 @@
+
+# WARNING
+
+This is an internal only package.
+
+We may regularly break changes in this package for our own internal uses. If you use anything from this
+package you do so at your own risk.
+
+(Also, this is a generated file. Do not change directly.)
