+++ packages/internal.foundry/src/public/OntologiesV2_OntologyObjectV2.ts
