+++ packages/internal.foundry/src/public/OntologiesV2_QueryType.ts
